Book: Assignments #time 94m

+++ fastapi/web/assignments/scrapping_eva.py
@@ -1,22 +1,36 @@
+"""
+>>> result[:5]  # doctest: +NORMALIZE_WHITESPACE
+[{'Number': 1, 'ISS': '', 'Mission': 'Voskhod 2', 'Participants': 'Alexi Leonov', 'Duration': '23m', 'Date': datetime.date(1965, 3, 18), 'Notes': 'First spacewalk'},
+ {'Number': 2, 'ISS': '', 'Mission': 'Gemini IV', 'Participants': 'Edward White, James McDivitt', 'Duration': '46m', 'Date': datetime.date(1966, 6, 3), 'Notes': None},
+ {'Number': 3, 'ISS': '', 'Mission': 'Gemini IX-A', 'Participants': 'Eugene Cernan, Thomas Stafford', 'Duration': '2h 8m', 'Date': datetime.date(1966, 6, 5), 'Notes': 'No egress by Stafford'},
+ {'Number': 4, 'ISS': '', 'Mission': 'Gemini X', 'Participants': 'Michael Collins, John Young', 'Duration': '50m', 'Date': datetime.date(1966, 7, 19), 'Notes': 'No egress by Young'},
+ {'Number': 5, 'ISS': '', 'Mission': 'Gemini X', 'Participants': 'Michael Collins, John Young', 'Duration': '40m', 'Date': datetime.date(1966, 7, 20), 'Notes': 'No egress by Young'}]
+"""
+
+from pathlib import Path
 import csv
 from datetime import datetime
-from pprint import pprint
 from bs4 import BeautifulSoup
 import requests
 
 
-URL = 'https://www.worldspaceflight.com/bios/eva/eva.php'
-FILE_HTML = r'eva.html'
-FILE_CSV = r'eva.csv'
+DATA = 'https://raw.githubusercontent.com/AstroMatt/book-python/master/_data/html/astro-eva.html'
+# DATA = 'https://www.worldspaceflight.com/bios/eva/eva.php'
+
+
+BASE_DIR = Path(__file__).parent.parent.parent.parent
+FILE_HTML = BASE_DIR.joinpath('_data/html/astro-eva.html')
+FILE_CSV = BASE_DIR.joinpath('_data/csv/astro-eva1.csv')
 
 
 # When you dump page for the first time, you can test parsing on local file
 # It saves bandwidth, and speeds you development
 # Then comment following `with` context manager
 
-with open(FILE_HTML, mode='w', encoding='utf-8') as file:
-    response = requests.get(URL)
-    file.write(response.text)
+if not FILE_HTML.exists():
+    with open(FILE_HTML, mode='w', encoding='utf-8') as file:
+        response = requests.get(DATA)
+        file.write(response.text)
 
 
 # Parser content below
@@ -73,12 +87,12 @@ def clean_Notes(text):
     return text
 
 
-records = []
+result = []
 
 for row in table.find_all('tr')[1:]:
     td = row.find_all('td')
 
-    records.append({
+    result.append({
         'Number': clean_Number(td[0].string),
         'ISS': clean_ISS(td[1].string),
         'Mission': clean_Mission(td[2].string),
@@ -92,7 +106,7 @@ for row in table.find_all('tr')[1:]:
 with open(FILE_CSV, mode='w', encoding='utf-8') as file:
     writer = csv.DictWriter(
         file,
-        fieldnames=records[0].keys(),
+        fieldnames=result[0].keys(),
         delimiter=',',
         quotechar='"',
         quoting=csv.QUOTE_ALL,
@@ -100,8 +114,5 @@ with open(FILE_CSV, mode='w', encoding='utf-8') as file:
 
     writer.writeheader()
 
-    for row in records:
+    for row in result:
         writer.writerow(row)
-
-
-pprint(records)
